feat!: drop old architecture
