+++ .changeset/breezy-camels-lead.md
@@ -1,0 +1,7 @@
+---
+'react-native-bottom-tabs': major
+'@bottom-tabs/expo-template': major
+'@bottom-tabs/react-navigation': major
+---
+
+feat!: drop old architecture
