+++ packages/react-native-bottom-tabs/android/build.gradle
@@ -15,7 +15,7 @@ buildscript {
 }
 
 def isNewArchitectureEnabled() {
-  return rootProject.hasProperty("newArchEnabled") && rootProject.getProperty("newArchEnabled") == "true"
+  return true
 }
 
 apply plugin: "com.android.library"
@@ -84,16 +84,11 @@ android {
 
   sourceSets {
     main {
-      if (isNewArchitectureEnabled()) {
-        java.srcDirs += [
-          "src/newarch",
-          // Codegen specs
-          "generated/java",
-          "generated/jni"
-        ]
-      } else {
-        java.srcDirs += ["src/oldarch"]
-      }
+      java.srcDirs += [
+        // Codegen specs
+        "generated/java",
+        "generated/jni"
+      ]
     }
   }
 }
