--- packages/react-native-bottom-tabs/ios/Events/TabLongPressedEvent.swift
@@ -1,38 +1,0 @@
-import React
-
-// RCTEvent is not defined for new arch.
-protocol RCTEvent {}
-
-@objcMembers
-public class TabLongPressEvent: NSObject, RCTEvent {
-  private var key: NSString
-  public var viewTag: NSNumber
-
-  public var eventName: String {
-    "onTabLongPress"
-  }
-
-  public init(reactTag: NSNumber, key: NSString) {
-    self.viewTag = reactTag
-    self.key = key
-    super.init()
-  }
-
-  public class func moduleDotMethod() -> String {
-    "RCTEventEmitter.receiveEvent"
-  }
-
-  public func canCoalesce() -> Bool {
-      false
-  }
-
-  public func arguments() -> [Any] {
-    [
-      viewTag,
-      RCTNormalizeInputEventName(eventName) ?? eventName,
-      [
-        "key": key
-      ]
-    ]
-  }
-}
